+++ scripts/run_benchmark/run_test_seqeracloud.sh
@@ -8,8 +8,8 @@ cd "$REPO_ROOT"
 
 set -e
 
-resources_test_s3=s3://openproblems-data/resources_test/denoising
-publish_dir_s3="s3://openproblems-nextflow/temp/results/denoising/$(date +%Y-%m-%d_%H-%M-%S)"
+resources_test_s3=s3://openproblems-data/resources_test/task_denoising
+publish_dir_s3="s3://openproblems-nextflow/temp/results/task_denoising/$(date +%Y-%m-%d_%H-%M-%S)"
 
 # write the parameters to file
 cat > /tmp/params.yaml << HERE
@@ -28,4 +28,4 @@ tw launch https://github.com/openproblems-bio/task_denoising.git \
   --compute-env 6TeIFgV5OY4pJCk8I0bfOh \
   --params-file /tmp/params.yaml \
   --config common/nextflow_helpers/labels_tw.config \
-  --labels denoising,test
+  --labels task_denoising,test