+++ src/data_processors/process_dataset/script.py
@@ -5,15 +5,16 @@ import numpy as np
 
 ## VIASH START
 par = {
-    'input': "resources_test/common/cxg_mouse_pancreas_atlas/dataset.h5ad",
-    'output_train': "train.h5ad",
-    'output_test': "test.h5ad",
+    'input': "resources/datasets/openproblems_v1/pancreas/log_cp10k/dataset.h5ad",
+    'output_train': "output/processed_datasets/train.h5ad",
+    'output_test': "output/processed_datasets/test.h5ad",
     'train_frac': 0.9,
-    'seed': 0
+    'seed': 0,
+    'n_obs_limit': 4000
 }
 meta = {
     "name": "process_dataset",
-    "resources_dir": "src/tasks/denoising/process_dataset"
+    "resources_dir": "src/data_processors/process_dataset"
 }
 ## VIASH END
 
@@ -29,27 +30,30 @@ adata = ad.read_h5ad(par["input"])
 
 # limit to max number of observations
 adata_output = adata.copy()
-if adata.n_obs > par["n_obs_limit"]:
-    print(">> Subsampling the observations", flush=True)
-    print(f">> Setting seed to {par['seed']}")
+
+if "batch" in adata.obs:
+    print(f">> Subsampling observations by largest batch", flush=True)
+    batch_counts = adata.obs.groupby('batch').size()
+    sorted_batches = batch_counts.sort_values(ascending=False)
+    selected_batch = sorted_batches.index[0]
+    adata_output = adata[adata.obs["batch"]==selected_batch,:].copy()
+
+if adata_output.n_obs > par["n_obs_limit"]:
+    print(f">> Randomly subsampling observations to {par['n_obs_limit']}", flush=True)
+    print(f">> Setting seed to {par['seed']}", flush=True)
     random.seed(par["seed"])
-    if "batch" not in adata.obs:
-        obs_filt = np.ones(dtype=np.bool_, shape=adata.n_obs)
-        obs_index = np.random.choice(np.where(obs_filt)[0], par["n_obs_limit"], replace=False)
-        adata_output = adata[obs_index].copy()
-    else:
-        batch_counts = adata.obs.groupby('batch').size()
-        filtered_batches = batch_counts[batch_counts <= par["n_obs_limit"]]
-        sorted_filtered_batches = filtered_batches.sort_values(ascending=False)
-        selected_batch = sorted_filtered_batches.index[0]
-        adata_output = adata[adata.obs["batch"]==selected_batch,:].copy()
+    obs_filt = np.ones(dtype=np.bool_, shape=adata_output.n_obs)
+    obs_index = np.random.choice(np.where(obs_filt)[0], par["n_obs_limit"], replace=False)
+    adata_output = adata_output[obs_index].copy()
         
 # remove all layers except for counts
+print(">> Remove all layers except for counts", flush=True)
 for key in list(adata_output.layers.keys()):
     if key != "counts":
         del adata_output.layers[key]
 
 # round counts and convert to int
+print(">> Round counts and convert to int", flush=True)
 counts = np.array(adata_output.layers["counts"]).round().astype(int)
 
 print(">> process and split data", flush=True)
@@ -65,6 +69,7 @@ X_train.eliminate_zeros()
 X_test.eliminate_zeros()
 
 # copy adata to train_set, test_set
+print(">> Create AnnData output objects", flush=True)
 output_train = ad.AnnData(
     layers={"counts": X_train},
     obs=adata_output.obs[[]],
@@ -83,6 +88,7 @@ output_test = ad.AnnData(
 output_test.uns["train_sum"] = X_train.sum()
 
 # Remove no cells that do not have enough reads
+print(">> Remove cells that do not have enough reads", flush=True)
 is_missing = np.array(X_train.sum(axis=0) == 0)
 
 output_train = output_train[:, ~is_missing.flatten()]